+++ src/Concerns/ManagesDataDefinitions.php
@@ -21,7 +21,6 @@ use Google\Cloud\Core\LongRunning\LongRunningOperation;
 use Google\Cloud\Spanner\Database;
 use RuntimeException;
 use function json_encode;
-use function trigger_deprecation;
 
 trait ManagesDataDefinitions
 {
@@ -29,28 +28,6 @@ trait ManagesDataDefinitions
      * @return Database
      */
     abstract public function getSpannerDatabase(): Database;
-
-    /**
-     * @deprecated use runDdlBatch() instead
-     * @param string $ddl
-     * @return LongRunningOperation
-     */
-    public function runDdl(string $ddl): LongRunningOperation
-    {
-        trigger_deprecation('colopl/laravel-spanner', '5.2', 'runDdl() is deprecated. Use runDdlBatch() instead.');
-        return $this->getSpannerDatabase()->updateDdl($ddl);
-    }
-
-    /**
-     * @deprecated use runDdlBatch() instead
-     * @param string[] $ddls
-     * @return LongRunningOperation
-     */
-    public function runDdls(array $ddls): LongRunningOperation
-    {
-        trigger_deprecation('colopl/laravel-spanner', '5.2', 'runDdls() is deprecated. Use runDdlBatch() instead.');
-        return $this->getSpannerDatabase()->updateDdlBatch($ddls);
-    }
 
     /**
      * @param list<string> $statements